WIP net_int_t & net_uint_t data types
Probably never gonna add this because it potentially breaks stuff in some scenarios.

+++ src/game/g_cmds.cpp
@@ -2038,11 +2038,7 @@ void G_Say(gentity_t *ent, gentity_t *target, int mode, qboolean encoded,
       color = COLOR_CYAN;
       break;
     case SAY_ADMIN:
-      Printer::logAdminLn(ETJump::stringFormat(
-          "adminchat: %i %s: %s", ent->s.number,
-          ETJump::sanitize(ent->client->pers.netname), chatText));
-      Com_sprintf(name, sizeof(name),
-                  "^A> ^7%s^7: ", ent->client->pers.netname);
+      Printer::logAdminLn(ETJump::stringFormat("adminchat: %i %s: %s", static_cast<int>(ent->s.number), ETJump::sanitize(ent->client->pers.netname), chatText)); Com_sprintf(name, sizeof(name), "^A> ^7%s^7: ", ent->client->pers.netname);
       color = COLOR_LTORANGE;
   }
